Add authentic ping. (#27)
* Add authentic ping.

I found a random youtube video online which had an actual
demonstration of an IBM selectric keyboard.

I then downloaded audacity and cut out the precise moment when
then ping occurred.  It was a little harsh, so I made the ping a
little dampened.

Enjoy!

* Add authentic selectric type sound.

Here is a sample of an actual person typing with the selectric keyboard.